requirements.txt
pillow==11.0.0
discord.py==2.4.0